security/requirements.txt
boto3==1.34.0
cryptography==42.0.5
python-dotenv==1.0.0
 